backend/requirements.txt
# Core Framework
Flask==3.0.0
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.36
alembic==1.13.0
# psycopg2-binary==2.9.9  # Commented: requires Visual C++ Build Tools on Windows + Python 3.13
# Alternative: Install PostgreSQL support only when needed in production (Linux/Docker)

# AI/Communication
openai==1.57.4
twilio==8.5.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
colorama==0.4.6
Werkzeug==3.0.0

# Task Queue - Requerido para recordatorios automáticos
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
